+++ pdfalyzer/util/cli_tools/argument_parsers.py
@@ -134,17 +134,19 @@ extract_text_parser.add_argument('--output-dir', '-out',
                                  help='write extracted text to .txt files of the same name as the PDF in this directory',
                                  type=DirValidator(allow_create=True))
 
-extract_text_parser.add_argument('--no-page-number-panels', '-n',
-                                 action='store_true',
+extract_text_parser.add_argument('--no-page-number-panels', '-n', action='store_true',
                                  help="don't print the PAGE 1, PAGE 2, etc. panels")
 
-extract_text_parser.add_argument('--page-range', '-r',
+extract_text_parser.add_argument('--page-range', '-pr',
                                  help=f"[PDFs only] {page_range_validator.HELP_MSG}",
                                  type=page_range_validator)
 
 extract_text_parser.add_argument('--print-as-parsed', '-p',
                                  action='store_true',
                                  help='print pages as they are parsed instead of waiting until parsing complete')
+
+extract_text_parser.add_argument('--recurse', '-r', action='store_true',
+                                 help=f"recursively traverse directory for .pdf files")
 
 
 def parse_text_extraction_args() -> Namespace:
@@ -158,7 +160,13 @@ def parse_text_extraction_args() -> Namespace:
             log.error(f"'{file_path}' is not a valid file or directory.")
             sys.exit(-1)
         elif file_path.is_dir():
-            args.files_to_process.extend(files_in_dir(file_path, 'pdf'))
+            if args.recurse:
+                pdf_files_in_dir = [f for f in file_path.glob('**/*.pdf')]
+                log.warning(f"Collected {len(pdf_files_in_dir)} PDF files via recursive traverse of {file_path}...")
+            else:
+                pdf_files_in_dir = files_in_dir(file_path, 'pdf')
+
+            args.files_to_process.extend(pdf_files_in_dir)
         else:
             args.files_to_process.append(file_path)